requirements.txt
flask
flask-cors
SQLAlchemy
gunicorn
Flask-SQLAlchemy
python-dotenv
